% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\docType{data}
\name{camp_rio_colors}
\alias{camp_rio_colors}
\title{Camp RIO Colors}
\format{
An object of class \code{list} of length 8.
}
\usage{
camp_rio_colors
}
\description{
Named list of colors from the [Camp RIO Brand Guidelines](https://ideapublicschoolsorg.sharepoint.com/:b:/s/ResearchandAnalysis/ETpoVdcAUQZCkXhI83eV6U0BHT2vFOcXUqAas326qEneXw?e=am5aeE)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: forestgreen, pond, sunriseyellow,
incarnadine, sunsetorange, midnight, riverblue, goldenhourbeige
}
\keyword{datasets}
