% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\docType{data}
\name{camp_rio_colors_2022}
\alias{camp_rio_colors_2022}
\title{Camp RIO Colors (2022)}
\format{
An object of class \code{list} of length 8.
}
\usage{
camp_rio_colors_2022
}
\description{
Named list of colors from the [Camp RIO Brand Guidelines (2022)](https://brandfolder.com/ideapublicschools)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: forestgreen, pond, sunriseyellow,
incarnadine, sunsetorange, midnight, riverblue, goldenhourbeige
}
\keyword{datasets}
