% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\docType{data}
\name{camp_rio_palettes}
\alias{camp_rio_palettes}
\title{Camp Rio Color Palettes}
\format{
An object of class \code{list} of length 5.
}
\usage{
camp_rio_palettes
}
\description{
A collection of color palettes based on the [IDEA Brand Guidelines](whttps://ideapublicschoolsorg.sharepoint.com/sites/MCE/Shared%20Documents/Forms/AllItems.aspx?id=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines%2Fbranding%5Fbrand%5Fguidelines%5F10x10%5F2019%2Epdf&parent=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available palettes is:
* qual
* div
* blueorange
* greenorange
* bluegray
* greengray
* orangegray
}
\examples{

# Make an x-y plot using the div palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) +
  scale_colour_idea() +
  theme_bw() +
  theme(aspect.ratio=1)

# Make a histogram using the qual palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_idea(palette="qual")
}
\keyword{datasets}
