% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\docType{data}
\name{camp_rio_palettes_2022}
\alias{camp_rio_palettes_2022}
\title{Camp RIO Color Palettes (2022)}
\format{
An object of class \code{list} of length 8.
}
\usage{
camp_rio_palettes_2022
}
\description{
A collection of color palettes based on the [Camp RIO Brand Guidelines (2022)](https://brandfolder.com/ideapublicschools)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available palettes is:
* qual
* div
* blueorange
* greenorange
* bluegray
* greengray
* orangegray
}
\examples{

# Make an x-y plot using the div palette
library(tidyverse)
df <- data.frame(x = rnorm(100, 0, 20),
          y = rnorm(100, 0, 20),
          cl = sample(letters[1:8], 100, replace=TRUE))
ggplot(df, aes(x, y, colour=cl, shape=cl)) +
  geom_point(size=4) +
  scale_colour_camp_rio() +
  theme_bw() +
  theme(aspect.ratio=1)

# Make a histogram using the qual palette
ggplot(df, aes(x, fill=cl)) + geom_histogram() +
  scale_fill_camp_rio(palette="qual")

}
\keyword{datasets}
