% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{idea_colors}
\alias{idea_colors}
\alias{idea_colors_2019}
\title{IDEA Public Schools Colors (2019)}
\format{
An object of class \code{list} of length 11.

An object of class \code{list} of length 11.
}
\usage{
idea_colors

idea_colors_2019
}
\description{
Named list of colors from the [IDEA Brand Guidelines (2019)](https://brandfolder.com/ideapublicschools)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: darkblue, blue, cyan, lime, magenta, melon,
yellow, coolgray, gray, lightgray, darkgray, red
}
\keyword{datasets}
