% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{idea_colors_2024}
\alias{idea_colors_2024}
\title{IDEA Public Schools Colors (2024)}
\format{
An object of class \code{list} of length 10.
}
\usage{
idea_colors_2024
}
\description{
Named list of colors from the [IDEA Brand Guidelines](https://brandfolder.com/ideapublicschools)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: blue, yellow, darkblue, lime, cyan, melon,
vermillion, coolgray, gray, lightgray.
}
\keyword{datasets}
