% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales_camp_rio.R
\name{scale_color_camp_rio}
\alias{scale_color_camp_rio}
\alias{scale_colour_camp_rio}
\title{Camp RIO color scale palette for ggplot2}
\usage{
scale_color_camp_rio(
  palette = "div",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)

scale_colour_camp_rio(
  palette = "div",
  discrete = TRUE,
  alpha = 1,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Choose from 'camp_rio_palettes' list}

\item{discrete}{whether to use a discrete color palette}

\item{alpha}{transparency}

\item{reverse}{logical, Reverse the order of the colours?}

\item{...}{additional arguments to passed to `scale_color_gradientn`}
}
\description{
Camp RIO color scale palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = factor(cyl))) +
  scale_colour_camp_rio(palette="qual")

ggplot(mtcars, aes(mpg, wt)) +
  geom_point(aes(colour = hp)) +
  scale_colour_camp_rio(palette="qual", discrete = FALSE)
ggplot(data = mpg) +
  geom_point(mapping = aes(x = displ, y = hwy, color = class)) +
  scale_colour_camp_rio(palette="bluegray")
}
