% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_idea}
\alias{scale_fill_idea}
\title{IDEA fill palette for ggplot2}
\usage{
scale_fill_idea(
  palette = "div",
  discrete = TRUE,
  alpha = 1,
  reverse = TRUE,
  year = 2024,
  ...
)
}
\arguments{
\item{palette}{Choose from 'idea_palettes' list}

\item{discrete}{whether to use a discrete colour palette}

\item{alpha}{The alpha transparency, a number in [0,1], see argument alpha in
\code{\link[grDevices]{hsv}}.}

\item{reverse}{If `TRUE``, the direction of the color ramp is reversed.}

\item{year}{Defaults to the most recent branding guidelines (year = 2024), but year = 2019 also available}

\item{...}{additional arguments to pass to scale_color_gradientn}
}
\description{
IDEA fill palette for ggplot2
}
\examples{
library(ggplot2)
ggplot(diamonds) + geom_bar(aes(x = cut, fill = clarity)) +
  scale_fill_idea()

}
