% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_idea_light}
\alias{theme_idea_light}
\title{Theme inspired by IDEA Brand Guidelines}
\usage{
theme_idea_light(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Theme inspired by the plots on
[IDEA Brand Guidelines](https://brandfolder.com/ideapublicschools).
}
\seealso{
Other themes idea: 
\code{\link{theme_idea_min}()}
}
\concept{themes idea}
