% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_idea_min}
\alias{theme_idea_min}
\title{Theme inspired by IDEA Brand Guidelines}
\usage{
theme_idea_min()
}
\description{
Theme inspired by the plots on
[IDEA Brand Guidelines](https://brandfolder.com/ideapublicschools).
}
\seealso{
Other themes idea: 
\code{\link{theme_idea_light}()}
}
\concept{themes idea}
