% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\docType{data}
\name{camp_rio_colors}
\alias{camp_rio_colors}
\title{Camp Rio Colors}
\format{
An object of class \code{list} of length 8.
}
\usage{
camp_rio_colors
}
\description{
Named list of colors from the [Camp Rio Brand Guidelines](https://ideapublicschoolsorg.sharepoint.com/sites/ResearchandAnalysis/Shared%20Documents/Forms/AllItems.aspx?id=%2Fsites%2FResearchandAnalysis%2FShared%20Documents%2FCamp%20Rio%2FCamp%20RIO%5FBranding%20Guidelines%5F2022%20%282%29%20%281%29%2Epdf&parent=%2Fsites%2FResearchandAnalysis%2FShared%20Documents%2FCamp%20Rio&p=true&ct=1702503731102&or=Teams%2DHL&ga=1)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: darkblue, blue, cyan, lime, magenta, melon,
yellow, coolgray, gray, lightgray, darkgray, red
}
\keyword{datasets}
