% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors_camp_rio.R
\name{camp_rio_palette_ramp}
\alias{camp_rio_palette_ramp}
\title{Camp Rio palettes with ramped colors}
\usage{
camp_rio_palette_ramp(palette = "div", alpha = 1, reverse = FALSE)
}
\arguments{
\item{palette}{Choose from `idea_palettes` list}

\item{alpha}{transparency from 0 (completely transparent) to 1 (completely opaque)}

\item{reverse}{If `TRUE``, the direction of the color ramp is reversed.}
}
\value{
a function that takes the number of colors needed as an argument
}
\description{
Camp Rio palettes with ramped colors
}
\examples{
library(scales)
show_col(camp_rio_palette_ramp()(10))

filled.contour(volcano,color.palette = camp_rio_palette_ramp(), asp=1)


}
