% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{idea_colors}
\alias{idea_colors}
\title{IDEA Public Schools Colors}
\format{
An object of class \code{list} of length 11.
}
\usage{
idea_colors
}
\description{
Named list of colors from the [IDEA Brand Guidelines](whttps://ideapublicschoolsorg.sharepoint.com/sites/MCE/Shared%20Documents/Forms/AllItems.aspx?id=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines%2Fbranding%5Fbrand%5Fguidelines%5F10x10%5F2019%2Epdf&parent=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines)
(you'll need to behind the firewall to see that guide).
}
\details{
The list of available colors is: darkblue, blue, cyan, lime, magenta, melon,
yellow, coolgray, gray, lightgray, darkgray, red
}
\keyword{datasets}
