% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_idea_light}
\alias{theme_idea_light}
\title{Theme inspired by IDEA Brand Guidelines}
\usage{
theme_idea_light(base_size = 12, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
Theme inspired by the plots on
[IDEA Brand Guidelines](https://ideapublicschoolsorg.sharepoint.com/sites/MCE/Shared%20Documents/Forms/AllItems.aspx?id=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines%2Fbranding%5Fbrand%5Fguidelines%5F10x10%5F2019%2Epdf&parent=%2Fsites%2FMCE%2FShared%20Documents%2FCreative%20Resources%2C%20Logos%2C%20and%20Mascots%2FBrand%20Guidelines).
}
\seealso{
Other themes idea: 
\code{\link{theme_idea_min}()}
}
\concept{themes idea}
